separete animation div
